Correction of errors, removal of unnecessary checks.
In the process of repairing # 6, a non-critical vulnerability was discovered in the implementation of the process of associating a user with a telegram account.
The fix affected the replacement of the reauthentication function to the correct one.
The check for the uniqueness of the user in the database was also removed, since this functionality was transferred to telegram_bot_user_mapping_add ()

Fix #6